+++ example/src/overlay.cpp
@@ -38,8 +38,8 @@ std::unique_ptr<Overlay> Overlay::FromFile(
 
 Overlay::Overlay(const gui::Context& context, Properties&& properties, gui::JSCallbackMap callbacks) :
   framework::Overlay(*context.renderer, context.driver.get(),
-                     properties.size.x, properties.size.y,
-                     properties.pos.x, properties.pos.y, properties.scale, true),
+                     (int)properties.size.x, (int)properties.size.y,
+                     (int)properties.pos.x, (int)properties.pos.y, properties.scale, true),
   callbacks_(callbacks) {
 }
